Fix problem of plugin exec executing every puppet run

+++ manifests/plugin.pp
@@ -64,6 +64,7 @@ define jenkins::plugin($version=0) {
     exec {
       "download-${name}" :
         command    => "rm -rf ${name} ${name}.* && wget --no-check-certificate ${base_url}${plugin}",
+        creates    => "${plugin_dir}/${plugin}",
         cwd        => $plugin_dir,
         require    => [File[$plugin_dir], Package['wget']],
         path       => ['/usr/bin', '/usr/sbin', '/bin'];
